Abandoned work on resolving unregistered services
The reason I've abandoned this is that it's very difficult to decorate
the built in service scope functionality. The issue here is that
ServiceProvider.CreateScope() - the native method not the extension
method - is `internal`.  This means I can't "do it the normal way" unless
I use an `IServiceScopeFactory`.  However, the whole point of this
logic is to register an alternative service scope factory, so I can't
inject "the default one".

Overall, this means I can't create scopes which "work the default way,
except also have my desired functionality".